HZN-865: Moved method into new Snmp4JUtils class.
